LORE-304 | Rename component
