+++ config/styleguide.config.json
@@ -45,8 +45,8 @@
       "IconSprite",
       "Input",
       "List",
-      "LocalNavigation",
       "Select",
+      "SimpleLocalNavigation",
       "Spinner",
       "Switch",
       "Switch.Item",
