+++ source/components/SimpleLocalNavigation/index.spec.js
@@ -1,14 +1,14 @@
 import { mount } from 'enzyme';
 import React from 'react';
 
-import LocalNavigation from './index';
+import SimpleLocalNavigation from './index';
 
-test('LocalNavigation renders', () => {
+test('SimpleLocalNavigation renders', () => {
     const component = mount(
-        <LocalNavigation className="foo">
+        <SimpleLocalNavigation className="foo">
             <span>Item 1</span>
             <span>Item 2</span>
-        </LocalNavigation>
+        </SimpleLocalNavigation>
     );
     expect(component).toMatchSnapshot();
 });
