+++ source/components/SimpleLocalNavigation/styles.scss
@@ -1,7 +1,7 @@
 @import "~design-system/dist/scss/wds-variables/index";
 @import "../FandomContentWell/styles.scss";
 
-.local-navigation {
+.simple-local-navigation {
 	background: #fff;
 	box-shadow: 0 2px 4px 0 rgba(0, 0, 0, .18);
